# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API permettant la gestion et la réservation de produits sous différentes formes, via des menus, des commandes. La mise en avant des produits ainsi que des promotions sont possible via celle-ci.",
    "version": "1.0.0",
    "title": "Burger API - Groupe 2",
    "license": {
      "name": "NodeJS",
      "url": "https://github.com/nodejs/node/blob/master/LICENSE"
    }
  },
  "host": "vincentguenin.ddns.net:8000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/user": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Ajouter un utilisateur",
        "description": "",
        "operationId": "user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Utilisateur ajouté à la base",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Succès de l'opération",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Paramètres invalides"
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
    "/user/signin": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Connexion d'un utilisateur",
        "description": "",
        "operationId": "user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Utilisateur connecté",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Succès de l'opération",
            "schema": {
              "$ref": "#/definitions/UserSignin"
            }
          },
          "400": {
            "description": "Paramètres invalides"
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
    "/user/logout": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Deconnexion d'un utilisateur",
        "description": "",
        "operationId": "user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Utilisateur Deconnecté",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Succès de l'opération",
            "schema": {
              "$ref": "#/definitions/UserSignin"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "email": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
    "UserSignin": {
      "type": "object",
      "properties": {
        "auth": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      }
    }
  }
}